Experimental Decoder
This decoder calculates offsets of the various struct fields, and directly sets the values.

Expected result was a significant performance increase, but it appears to slightly worse currently, because reflect internally turns it's unsafe.Pointer into a uintptr when we call UnsafeAddr()

Could perhaps be more performant if we compiled a MessageDescriptor to Struct path, which skips the string-based lookup, and instead just iterate an array of simple instructions.